scheduler_service/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DB_USER = os.getenv("DB_USER")
DB_PASSWORD = os.getenv("DB_PASSWORD")
DB_HOST = os.getenv("DB_HOST")
DB_PORT = os.getenv("DB_PORT")
DB_NAME = os.getenv("DB_NAME")

DB_URL = f"postgresql://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
ROUTER_SERVICE_URL = os.getenv("ROUTER_SERVICE_URL", "http://router:8000")

MAX_DATA_DATE = os.getenv("MAX_DATA_DATE", "2017-05-10")