scheduler_service/requirements.txt
fastapi
apscheduler
httpx
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary